Chunked the giant index.tsx file into smaller components files
